+++ TelemetryPoint.h
@@ -3,9 +3,9 @@
 
 struct TelemetryPoint
 {
-    int x;
-    int y;
-    int quality;
+    int16_t x;
+    int16_t y;
+    int16_t quality;
     float distance;
     float angle;
 };
